+++ utils/date.ts
@@ -1,0 +1,7 @@
+export const isSameDay = (lhs: Date, rhs: Date) => {
+  return (
+    lhs.getFullYear() === rhs.getFullYear() &&
+    lhs.getMonth() === rhs.getMonth() &&
+    lhs.getDate() === rhs.getDate()
+  );
+};
